requirements.txt
psutil
requests
pyspark==3.2.0
kafka-python==2.0.2
confluent-kafka==1.7.0
schedule